feat(events) : put helper functions into a new class
